eyeul: Expose "HTC Gesture sensor" as type ID 0x10001
HTC Desire Eye exposes a custom sensor "HTC Gesture sensor" with type 27,
which is used for SENSOR_TYPE_DEVICE_ORIENTATION sensors. This is now
an issue in Pie, since it is a one-shot sensor and cannot be
registered by the WindowOrientationListener.

Workaround the issue by replacing all the occurrences of 0x1B
(the stock value of 'type', which is a member of 'sensor_t' struct)
to 0x10001 in the sensors HAL.

Change-Id: I99ac93ec485d7c349eb705a4e953e5410d1ae025